migrated case logic into cases.py
